* master:
  Fix x64 warnings
  Fix x64 warnings
  Fix x64 warnings
  Cleanup
  Fix warnings
  Fix x64 warnings
  Fix x64 warnings
  Fix x64 warnings
  Fix x64 warnings
  WebDAVFileSystem.cpp: plaintext_prompt_helper: Remove loop
  MFC: Fix compilation
  afxglobals: Cleanup
  MFC: Cleanup
  MFC: Cleanup
  MFC: Remove CAtlTransactionManager; remove atlfile.h
  MFC: Cleanup
  Bug 948: Session settings TunnelHostKey is not working

Conflicts:
	src/NetBox/FarConfiguration.cpp
	src/NetBox/FarConfiguration.h
	src/NetBox/FarDialog.cpp
	src/NetBox/FarPlugin.cpp
	src/NetBox/FarPlugin.h
	src/NetBox/NetBox.cpp

+++ src/NetBox/FarConfiguration.cpp
@@ -270,7 +270,7 @@ void TFarConfiguration::CacheFarSettings()
   FFarConfirmations = GetConfirmationsSettings();
 }
 //---------------------------------------------------------------------------
-__int64 TFarConfiguration::FarConfirmations()
+intptr_t TFarConfiguration::FarConfirmations()
 {
   if (GetCurrentThreadId() == GetPlugin()->GetFarThread())
   {
