+++ src/NetBox/FarConfiguration.h
@@ -92,7 +92,7 @@ private:
 
 private:
   TCustomFarPlugin * FFarPlugin;
-  int FFarConfirmations;
+  intptr_t FFarConfirmations;
   bool FConfirmOverwritingOverride;
   bool FConfirmSynchronizedBrowsing;
   bool FForceInheritance;
@@ -121,7 +121,7 @@ private:
 
   TBookmarks * FBookmarks;
 
-  __int64 FarConfirmations();
+  intptr_t FarConfirmations();
 };
 //---------------------------------------------------------------------------
 #define FarConfiguration (static_cast<TFarConfiguration *>(Configuration))
